Split fixtures in the GiftCard module
