+++ saleor/giftcard/tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .giftcard import *  # noqa: F403
+from .giftcard_event import *  # noqa: F403
+from .giftcard_tag import *  # noqa: F403
