+++ saleor/graphql/giftcard/tests/benchmark/__init__.py
@@ -1,0 +1,1 @@
+GIFT_CARD_COUNT_IN_BENCHMARKS = 20
