+++ saleor/tests/fixtures.py
@@ -66,8 +66,6 @@ from ..discount.models import (
     VoucherCustomer,
     VoucherTranslation,
 )
-from ..giftcard import GiftCardEvents
-from ..giftcard.models import GiftCard, GiftCardEvent, GiftCardTag
 from ..payment import ChargeStatus, TransactionKind
 from ..payment.interface import AddressData, GatewayConfig, GatewayResponse, PaymentData
 from ..payment.models import Payment, TransactionEvent, TransactionItem
@@ -2099,138 +2097,6 @@ def checkout_line_with_reserved_preorder_item(
     )
 
     return checkout_line
-
-
-@pytest.fixture
-def gift_card_tag_list(db):
-    tags = [GiftCardTag(name=f"test-tag-{i}") for i in range(5)]
-    return GiftCardTag.objects.bulk_create(tags)
-
-
-@pytest.fixture
-def gift_card(customer_user):
-    gift_card = GiftCard.objects.create(
-        code="never_expiry",
-        created_by=customer_user,
-        created_by_email=customer_user.email,
-        initial_balance=Money(10, "USD"),
-        current_balance=Money(10, "USD"),
-    )
-    tag, _ = GiftCardTag.objects.get_or_create(name="test-tag")
-    gift_card.tags.add(tag)
-    return gift_card
-
-
-@pytest.fixture
-def gift_card_with_metadata(customer_user):
-    return GiftCard.objects.create(
-        code="card_with_meta",
-        created_by=customer_user,
-        created_by_email=customer_user.email,
-        initial_balance=Money(10, "USD"),
-        current_balance=Money(10, "USD"),
-        metadata={"test": "value"},
-    )
-
-
-@pytest.fixture
-def gift_card_expiry_date(customer_user):
-    gift_card = GiftCard.objects.create(
-        code="expiry_date",
-        created_by=customer_user,
-        created_by_email=customer_user.email,
-        initial_balance=Money(20, "USD"),
-        current_balance=Money(20, "USD"),
-        expiry_date=datetime.datetime.now(tz=datetime.UTC).date()
-        + datetime.timedelta(days=100),
-    )
-    tag = GiftCardTag.objects.create(name="another-tag")
-    gift_card.tags.add(tag)
-    return gift_card
-
-
-@pytest.fixture
-def gift_card_used(staff_user, customer_user):
-    gift_card = GiftCard.objects.create(
-        code="giftcard_used",
-        created_by=staff_user,
-        used_by=customer_user,
-        created_by_email=staff_user.email,
-        used_by_email=customer_user.email,
-        initial_balance=Money(100, "USD"),
-        current_balance=Money(80, "USD"),
-    )
-    tag = GiftCardTag.objects.create(name="tag")
-    gift_card.tags.add(tag)
-    return gift_card
-
-
-@pytest.fixture
-def gift_card_created_by_staff(staff_user):
-    gift_card = GiftCard.objects.create(
-        code="created_by_staff",
-        created_by=staff_user,
-        created_by_email=staff_user.email,
-        initial_balance=Money(10, "USD"),
-        current_balance=Money(10, "USD"),
-    )
-    tag, _ = GiftCardTag.objects.get_or_create(name="test-tag")
-    gift_card.tags.add(tag)
-    return gift_card
-
-
-@pytest.fixture
-def gift_card_event(gift_card, order, app, staff_user):
-    parameters = {
-        "message": "test message",
-        "email": "testemail@email.com",
-        "tags": ["test tag"],
-        "old_tags": ["test old tag"],
-        "balance": {
-            "currency": "USD",
-            "initial_balance": 10,
-            "old_initial_balance": 20,
-            "current_balance": 10,
-            "old_current_balance": 5,
-        },
-        "expiry_date": datetime.date(2050, 1, 1),
-        "old_expiry_date": datetime.date(2010, 1, 1),
-    }
-    return GiftCardEvent.objects.create(
-        user=staff_user,
-        app=app,
-        gift_card=gift_card,
-        order=order,
-        type=GiftCardEvents.UPDATED,
-        parameters=parameters,
-        date=timezone.now() + datetime.timedelta(days=10),
-    )
-
-
-@pytest.fixture
-def gift_card_list():
-    gift_cards = list(
-        GiftCard.objects.bulk_create(
-            [
-                GiftCard(
-                    code="code-test-1",
-                    initial_balance=Money(10, "USD"),
-                    current_balance=Money(10, "USD"),
-                ),
-                GiftCard(
-                    code="code-test-2",
-                    initial_balance=Money(10, "USD"),
-                    current_balance=Money(10, "USD"),
-                ),
-                GiftCard(
-                    code="code-test-3",
-                    initial_balance=Money(10, "USD"),
-                    current_balance=Money(10, "USD"),
-                ),
-            ]
-        )
-    )
-    return gift_cards
 
 
 @pytest.fixture
